Moved the validation functions from command_validator.rb to beefy.rb so the Beefy class is self contained and can be run from the single Beefy.rb file as well as from simulator.rb.
Removed puts statements when the Beefy robot is ignoring invalid MOVE, LEFT, RIGHT and REPORT commands (for when no placement exists)

Added validation to is_placement_valid? method to check if the x and y coordinates are integers or string representations of integers '1', '2',... I've assumed that input like 'one' and 'two' are illegal

Added report method that aliasses Beefy.to_s method so the same commands work as in the simulator to what is in the requirements (ie. PLACE, MOVE, LEFT, RIGHT and REPORT in commands.txt are equivalent to Beefy methods place, move, left, right and report).

Deleted command_validator.rb